requirements.txt
joblib
fastapi
uvicorn
pydantic
wandb
catboost
xgboost
imbalanced-learn
python-dotenv
scikit-learn
seaborn
matplotlib
pandas
numpy